feat: enable graceful exits for DeviceManager and improve logging control
