+++ CHANGELOG.md
@@ -22,6 +22,10 @@ Things to be included in the next release go here.
 
 - Updated the `PIControl.query_binary()` method to include PyVISA-compatible arguments, enabling users to fully utilize the complete functionality of the PyVISA method.
 
+### Fixed
+
+- Updated the `DeviceManager` to gracefully exit if the close method is not called in any scripts.
+
 ### Removed
 
 - Python 3.8 support has been removed from the package. The minimum supported version is now Python 3.9.
