+++ src/tm_devices/helpers/__init__.py
@@ -25,6 +25,7 @@ from tm_devices.helpers.functions import (
     check_visa_connection,
     create_visa_connection,
     detect_visa_resource_expression,
+    disable_all_loggers,
     get_model_series,
     get_version,
     get_visa_backend,
@@ -73,6 +74,7 @@ __all__ = [
     "configure_logging",
     "create_visa_connection",
     "detect_visa_resource_expression",
+    "disable_all_loggers",
     "get_model_series",
     "get_version",
     "get_visa_backend",
